Finished blackjack view
modified last details of the blackjack view, such as the insertion of a pop-up window for exiting the game. Also fixed the player's balance, now when the game ends he will exit directly from the game

+++ src/main/java/ludomania/handler/BlackJackHandler.java
@@ -8,7 +8,7 @@ public interface BlackJackHandler {
 
     void handleStand();
 
-    void handlePlaceBet(int amount);
+    void handlePlaceBet(double amount);
 
     void handleStartGame();
